% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function}
\alias{dummy_function}
\title{Dummy function}
\usage{
dummy_function(...)
}
\arguments{
\item{...}{dummy dots}
}
\value{
'dummy'
}
\description{
Dummy function
}
\examples{
dummy_function()

}
