% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_multiplier}
\alias{dummy_multiplier}
\title{Another dummy function}
\usage{
dummy_multiplier(n = 2)
}
\arguments{
\item{n}{integer}
}
\value{
character 'dummy' repeated n times
}
\description{
This is another function very very dummy
}
\details{
With this function, we not only get 'dummy', but 'dummy' multiplied by n times
}
\examples{
dummy_multiplier(3)

}
